requirements.txt
# DINOv2 Offroad Segmentation Ensemble
# Duality AI Hackathon Dependencies

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Scientific
numpy>=1.24.0
scipy>=1.11.0

# Visualization & Metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
tqdm>=4.65.0

# Data
pandas>=2.0.0
pyyaml>=6.0.0

# Utilities
pathlib
warnings

# For Kaggle/notebook environments
# jupyter (optional, for local notebook testing)
# ipykernel (optional)
